\name{poilog-package}
\alias{poilog-package}
\alias{poilog}
\docType{package}
\title{
Poisson lognormal and bivariate Poisson lognormal distribution
}
\description{
Functions for obtaining the density, random deviates and maximum likelihood 
estimates of the Poisson lognormal distribution and the bivariate Poisson 
lognormal distribution.
}
\details{
\tabular{ll}{
Package: \tab poilog\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2007-12-03\cr
License: \tab GPL-3\cr
}
\code{\link{dpoilog}} returns the density, \code{\link{rpoilog}} returns random 
deviates and \code{\link{poilogMLE}} performs maximum likelihood estimation of
parameters for the Poisson lognormal distribution.

\code{\link{dbipoilog}} returns the density, \code{\link{rbipoilog}} returns random 
deviates and \code{\link{bipoilogMLE}} performs maximum likelihood estimation of
parameters for the bivariate Poisson lognormal distribution.

}
\author{
Vidar \enc{Grøtan}{Grotan} and Steinar Engen

Maintainer: 
Vidar \enc{Grøtan}{Grotan} \email{vidar.grotan@bio.ntnu.no}
}
\references{
Bulmer, M. G. 1974. On fitting the Poisson lognormal distribution to species abundance data. Biometrics \bold{30}, 651-660.\cr
Engen, S., R. Lande, T. Walla and P. J. DeVries. 2002. Analyzing spatial structure of communities using the two-dimensional 
            Poisson lognormal species abundance model. American Naturalist \bold{160}, 60-73.
}

\keyword{ package }
\encoding{latin1}
